% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_model.R
\name{random_model}
\alias{random_model}
\title{A function to create randomized parameters for models that are of particular
importance when considering potential simulated timeseries.}
\usage{
random_model(nf, option = "gaussian")
}
\arguments{
\item{nf:}{the number of features.}

\item{option:}{the model to sample from. Options are 'gaussian'.}
}
\value{
model: a model to provide parameters for the particular distribution chosen.
\itemize{
 \item gaussian model:
   \itemize{
     \item model$type = 'gaussian'
     \item model$mean[nf]: a mean vector for the timeseries to be generated.
     \item model$cov[nf, nf]: a covariance matrix for the timeseries.
   }
}
}
\description{
A function to create randomized parameters for models that are of particular
importance when considering potential simulated timeseries.
}
\examples{
test <- random_model(option='gaussian')
}

