% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_random_population.R
\name{sample_random_population}
\alias{sample_random_population}
\title{A function to randomly sample a population of timeseries from a user-defined model type.}
\usage{
sample_random_population(numsubs, numsessions, nt, nf, option = "gaussian",
  noise = FALSE, simplify = FALSE)
}
\arguments{
\item{numsubs:}{the number of subjects.}

\item{numsessions:}{the number of sessions per subject.}

\item{nt:}{the number of timesteps.}

\item{nf:}{the number of features.}

\item{options:}{the model to sample. Options are:
\itemize{
   \item 'gaussian': a gaussian random model.
}}

\item{noise:}{a parameter whether to add zero-mean noise. Provide the sigma (noise=0 is to add no noise).}

\item{simplify:}{a parameter specifying whether to simplify the output. Options are FALSE and 'array'.}
}
\value{
ts: the output population timeseries. Can be formatted as:
\itemize{
   \item if simplify=FALSE:
   \itemize{
     \item ts[[sub]][[session]]$signal: the signal for a particular subject. Dimensions are [timesteps, features].
     \item ts[[sub]][[session]]$model: the model used for a particular subject.
   }
   \item if simplify='array':
   \itemize{
     \item ts[timestep, feature, subject, session]
   }
}
}
\description{
A function to randomly sample a population of timeseries from a user-defined model type.
}
\examples{
ts <- sample_random_population(4, 2, 100, 1, option='gaussian')
}

