% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_random_timeseries.R
\name{sample_random_timeseries}
\alias{sample_random_timeseries}
\title{A function to randomly sample a single random timeseries}
\usage{
sample_random_timeseries(nt, nf, option = "gaussian", model = NULL,
  noise = FALSE, simplify = FALSE)
}
\arguments{
\item{nt:}{the number of timesteps.}

\item{nf:}{the number of features.}

\item{options:}{the model to sample. Options are:
\itemize{
   \item 'gaussian': a gaussian random model.
   \item if model is specified, this option is irrelevant.
}}

\item{model:}{a model to provide parameters for the particular distribution chosen.
\itemize{
 \item gaussian model:
   \itemize{
     \item model$type = 'gaussian'
     \item model$mean[nf]: a mean vector for the timeseries to be generated.
     \item model$cov[nf, nf]: a covariance matrix for the timeseries.
   }
}}

\item{noise:}{a parameter whether to add zero-mean noise. Provide the sigma (noise=FALSE is to add no noise).}

\item{simplify:}{A parameter whether to simplify the output.
\itemize{
 \item if simplify=='TRUE'array': returns only the signal.
 \item if simplify==FALSE: returns a list of the signal and the model.
}}
}
\value{
signal[nt, nf]: A random timeseries.

model: the model chosen. Only returned if simplify=FALSE.
}
\description{
A function to randomly sample a single random timeseries
}
\examples{
test <- sample_random_timeseries(100, 10, options='gaussian')
dim(test)
}

